gif support
Added shader for gifs for low-budget "particle"-like effects. These can be added to billboards to polish the "despawn" events. Need to find better gifs (for some reason the parser that was used for the shader can't read the "poof.gif" image. Pushing now to test performance of shader on mobile