Fixing 6dof errors w/ grabbing clones
I *think* both static body collisions and the "sphere-collider" component  both emit a "hit" event from their respective strategies of detecting collisions... the "sphere-collider" filters objects based on an "objects" property, but the physics system collisions do not... I think the old demo added conditions for "grabbing" objects based on a class, so a modification to the "grab" class in "a-frame-physics-extras" may need to be used (one that is updated). Currently spawners do appear when VR mode is entered with a 6dof control scheme, and when the hands (which also appear for only 6dof control schemes) try to "grab" the spawner a new Ram's Horn is spawned (the orientation is currently wrong), but these are not "grabbed". I *think* the grab component has already registered a "grabbed" element, the spawner collider... this may be incorrect, it kind of warrants more debugging.